Tools: Moved themes into Extensions folder.
