+++ js/Extensions/Themes/Avocado.js
@@ -12,34 +12,28 @@
  *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
  *
  * */
-
 'use strict';
-
-import Highcharts from '../Core/Globals.js';
-import U from '../Core/Utilities.js';
-const { setOptions } = U;
-
+import Highcharts from '../../Core/Globals.js';
+import U from '../../Core/Utilities.js';
+var setOptions = U.setOptions;
 Highcharts.theme = {
-    colors: ['#FDD089', '#FF7F79', '#A0446E', '#251535'],
-
+    colors: ['#F3E796', '#95C471', '#35729E', '#251735'],
     colorAxis: {
-        maxColor: '#60042E',
-        minColor: '#FDD089'
+        maxColor: '#05426E',
+        minColor: '#F3E796'
     },
-
     plotOptions: {
         map: {
-            nullColor: '#fefefc'
+            nullColor: '#FCFEFE'
         }
-    } as Highcharts.PlotOptions,
-
+    },
     navigator: {
+        maskFill: 'rgba(170, 205, 170, 0.5)',
         series: {
-            color: '#FF7F79',
-            lineColor: '#A0446E'
+            color: '#95C471',
+            lineColor: '#35729E'
         }
     }
 };
-
 // Apply the theme
 setOptions(Highcharts.theme);
