+++ js/Extensions/Themes/Grid.js
@@ -9,13 +9,10 @@
  *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
  *
  * */
-
 'use strict';
-
-import Highcharts from '../Core/Globals.js';
-import U from '../Core/Utilities.js';
-const { setOptions } = U;
-
+import Highcharts from '../../Core/Globals.js';
+import U from '../../Core/Utilities.js';
+var setOptions = U.setOptions;
 Highcharts.theme = {
     colors: ['#058DC7', '#50B432', '#ED561B', '#DDDF00', '#24CBE5', '#64E572',
         '#FF9655', '#FFF263', '#6AF9C4'],
@@ -60,7 +57,6 @@ Highcharts.theme = {
                 fontWeight: 'bold',
                 fontSize: '12px',
                 fontFamily: 'Trebuchet MS, Verdana, sans-serif'
-
             }
         }
     },
@@ -89,7 +85,6 @@ Highcharts.theme = {
         itemStyle: {
             font: '9pt Trebuchet MS, Verdana, sans-serif',
             color: 'black'
-
         },
         itemHoverStyle: {
             color: '#039'
@@ -103,7 +98,6 @@ Highcharts.theme = {
             color: '#99b'
         }
     },
-
     navigation: {
         buttonOptions: {
             theme: {
@@ -112,6 +106,5 @@ Highcharts.theme = {
         }
     }
 };
-
 // Apply the theme
 setOptions(Highcharts.theme);
