+++ js/Extensions/Themes/HighContrastDark.js
@@ -1,64 +1,55 @@
 /* *
  *
- *  (c) 2010-2020 Torstein Honsi
+ *  (c) 2010-2020 Highsoft AS
+ *
+ *  Author: Øystein Moseng
  *
  *  License: www.highcharts.com/license
  *
- *  Dark theme for Highcharts JS
+ *  Accessible high-contrast dark theme for Highcharts. Specifically tailored
+ *  towards 3:1 contrast against black/off-black backgrounds. Neighboring
+ *  colors are tested for color blindness.
  *
  *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
  *
  * */
-
 'use strict';
-
-/* global document */
-
-// Load the fonts
-import Highcharts from '../Core/Globals.js';
-import U from '../Core/Utilities.js';
-const { setOptions } = U;
-
-Highcharts.createElement('link', {
-    href: 'https://fonts.googleapis.com/css?family=Unica+One',
-    rel: 'stylesheet',
-    type: 'text/css'
-}, null as any, document.getElementsByTagName('head')[0]);
-
+import Highcharts from '../../Core/Globals.js';
+import U from '../../Core/Utilities.js';
+var setOptions = U.setOptions;
+var textBright = '#F0F0F3';
 Highcharts.theme = {
-    colors: ['#2b908f', '#90ee7e', '#f45b5b', '#7798BF', '#aaeeee', '#ff0066',
-        '#eeaaee', '#55BF3B', '#DF5353', '#7798BF', '#aaeeee'],
+    colors: [
+        '#a6f0ff',
+        '#70d49e',
+        '#e898a5',
+        '#007faa',
+        '#f9db72',
+        '#f45b5b',
+        '#1e824c',
+        '#e7934c',
+        '#dadfe1',
+        '#a0618b'
+    ],
     chart: {
-        backgroundColor: {
-            linearGradient: { x1: 0, y1: 0, x2: 1, y2: 1 },
-            stops: [
-                [0, '#2a2a2b'],
-                [1, '#3e3e40']
-            ]
-        },
-        style: {
-            fontFamily: '\'Unica One\', sans-serif'
-        },
+        backgroundColor: '#1f1f20',
         plotBorderColor: '#606063'
     },
     title: {
         style: {
-            color: '#E0E0E3',
-            textTransform: 'uppercase',
-            fontSize: '20px'
+            color: textBright
         }
     },
     subtitle: {
         style: {
-            color: '#E0E0E3',
-            textTransform: 'uppercase'
+            color: textBright
         }
     },
     xAxis: {
         gridLineColor: '#707073',
         labels: {
             style: {
-                color: '#E0E0E3'
+                color: textBright
             }
         },
         lineColor: '#707073',
@@ -66,8 +57,7 @@ Highcharts.theme = {
         tickColor: '#707073',
         title: {
             style: {
-                color: '#A0A0A3'
-
+                color: textBright
             }
         }
     },
@@ -75,32 +65,28 @@ Highcharts.theme = {
         gridLineColor: '#707073',
         labels: {
             style: {
-                color: '#E0E0E3'
+                color: textBright
             }
         },
         lineColor: '#707073',
         minorGridLineColor: '#505053',
         tickColor: '#707073',
-        tickWidth: 1,
         title: {
             style: {
-                color: '#A0A0A3'
+                color: textBright
             }
         }
     },
     tooltip: {
         backgroundColor: 'rgba(0, 0, 0, 0.85)',
         style: {
-            color: '#F0F0F0'
+            color: textBright
         }
     },
     plotOptions: {
         series: {
             dataLabels: {
-                color: '#F0F0F3',
-                style: {
-                    fontSize: '13px'
-                }
+                color: textBright
             },
             marker: {
                 lineColor: '#333'
@@ -114,12 +100,15 @@ Highcharts.theme = {
         },
         errorbar: {
             color: 'white'
+        },
+        map: {
+            nullColor: '#353535'
         }
-    } as Highcharts.PlotOptions,
+    },
     legend: {
-        backgroundColor: 'rgba(0, 0, 0, 0.5)',
+        backgroundColor: 'transparent',
         itemStyle: {
-            color: '#E0E0E3'
+            color: textBright
         },
         itemHoverStyle: {
             color: '#FFF'
@@ -129,13 +118,13 @@ Highcharts.theme = {
         },
         title: {
             style: {
-                color: '#C0C0C0'
+                color: '#D0D0D0'
             }
         }
     },
     credits: {
         style: {
-            color: '#666'
+            color: textBright
         }
     },
     labels: {
@@ -143,16 +132,14 @@ Highcharts.theme = {
             color: '#707073'
         }
     },
-
     drilldown: {
         activeAxisLabelStyle: {
-            color: '#F0F0F3'
+            color: textBright
         },
         activeDataLabelStyle: {
-            color: '#F0F0F3'
+            color: textBright
         }
     },
-
     navigation: {
         buttonOptions: {
             symbolStroke: '#DDDDDD',
@@ -161,28 +148,26 @@ Highcharts.theme = {
             }
         }
     },
-
-    // scroll charts
     rangeSelector: {
         buttonTheme: {
             fill: '#505053',
             stroke: '#000000',
             style: {
-                color: '#CCC'
+                color: '#eee'
             },
             states: {
                 hover: {
                     fill: '#707073',
                     stroke: '#000000',
                     style: {
-                        color: 'white'
+                        color: textBright
                     }
                 },
                 select: {
-                    fill: '#000003',
-                    stroke: '#000000',
+                    fill: '#303030',
+                    stroke: '#101010',
                     style: {
-                        color: 'white'
+                        color: textBright
                     }
                 }
             }
@@ -190,20 +175,19 @@ Highcharts.theme = {
         inputBoxBorderColor: '#505053',
         inputStyle: {
             backgroundColor: '#333',
-            color: 'silver'
+            color: textBright
         },
         labelStyle: {
-            color: 'silver'
+            color: textBright
         }
     },
-
     navigator: {
         handles: {
             backgroundColor: '#666',
             borderColor: '#AAA'
         },
         outlineColor: '#CCC',
-        maskFill: 'rgba(255,255,255,0.1)',
+        maskFill: 'rgba(180,180,255,0.2)',
         series: {
             color: '#7798BF',
             lineColor: '#A6C7ED'
@@ -212,7 +196,6 @@ Highcharts.theme = {
             gridLineColor: '#505053'
         }
     },
-
     scrollbar: {
         barBackgroundColor: '#808083',
         barBorderColor: '#808083',
@@ -224,6 +207,5 @@ Highcharts.theme = {
         trackBorderColor: '#404043'
     }
 };
-
 // Apply the theme
 setOptions(Highcharts.theme);
