+++ ts/Extensions/Themes/Avocado.ts
@@ -12,21 +12,26 @@
  *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
  *
  * */
+
 'use strict';
-import Highcharts from '../Core/Globals.js';
-import U from '../Core/Utilities.js';
-var setOptions = U.setOptions;
+import Highcharts from '../../Core/Globals.js';
+import U from '../../Core/Utilities.js';
+const { setOptions } = U;
+
 Highcharts.theme = {
     colors: ['#F3E796', '#95C471', '#35729E', '#251735'],
+
     colorAxis: {
         maxColor: '#05426E',
         minColor: '#F3E796'
     },
+
     plotOptions: {
         map: {
             nullColor: '#FCFEFE'
         }
-    },
+    } as Highcharts.PlotOptions,
+
     navigator: {
         maskFill: 'rgba(170, 205, 170, 0.5)',
         series: {
@@ -35,5 +40,6 @@ Highcharts.theme = {
         }
     }
 };
+
 // Apply the theme
 setOptions(Highcharts.theme);
