+++ ts/Extensions/Themes/DarkBlue.ts
@@ -4,127 +4,99 @@
  *
  *  License: www.highcharts.com/license
  *
- *  Gray theme for Highcharts JS
+ *  Dark blue theme for Highcharts JS
  *
  *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
  *
  * */
+
 'use strict';
-import Highcharts from '../Core/Globals.js';
-import U from '../Core/Utilities.js';
-var setOptions = U.setOptions;
+import Highcharts from '../../Core/Globals.js';
+import U from '../../Core/Utilities.js';
+const { setOptions } = U;
+
 Highcharts.theme = {
-    colors: ['#DDDF0D', '#7798BF', '#55BF3B', '#DF5353', '#aaeeee',
+    colors: ['#DDDF0D', '#55BF3B', '#DF5353', '#7798BF', '#aaeeee',
         '#ff0066', '#eeaaee', '#55BF3B', '#DF5353', '#7798BF', '#aaeeee'],
     chart: {
         backgroundColor: {
-            linearGradient: { x1: 0, y1: 0, x2: 0, y2: 1 },
+            linearGradient: { x1: 0, y1: 0, x2: 1, y2: 1 },
             stops: [
-                [0, 'rgb(96, 96, 96)'],
-                [1, 'rgb(16, 16, 16)']
+                [0, 'rgb(48, 48, 96)'],
+                [1, 'rgb(0, 0, 0)']
             ]
         },
-        borderWidth: 0,
-        borderRadius: 0,
-        plotBackgroundColor: null,
-        plotShadow: false,
-        plotBorderWidth: 0
+        borderColor: '#000000',
+        borderWidth: 2,
+        className: 'dark-container',
+        plotBackgroundColor: 'rgba(255, 255, 255, .1)',
+        plotBorderColor: '#CCCCCC',
+        plotBorderWidth: 1
     },
     title: {
         style: {
-            color: '#FFF',
-            font: '16px Lucida Grande, Lucida Sans Unicode,' +
-                ' Verdana, Arial, Helvetica, sans-serif'
+            color: '#C0C0C0',
+            font: 'bold 16px "Trebuchet MS", Verdana, sans-serif'
         }
     },
     subtitle: {
         style: {
-            color: '#DDD',
-            font: '12px Lucida Grande, Lucida Sans Unicode,' +
-                ' Verdana, Arial, Helvetica, sans-serif'
+            color: '#666666',
+            font: 'bold 12px "Trebuchet MS", Verdana, sans-serif'
         }
     },
     xAxis: {
-        gridLineWidth: 0,
-        lineColor: '#999',
-        tickColor: '#999',
+        gridLineColor: '#333333',
+        gridLineWidth: 1,
         labels: {
             style: {
-                color: '#999',
-                fontWeight: 'bold'
+                color: '#A0A0A0'
             }
         },
+        lineColor: '#A0A0A0',
+        tickColor: '#A0A0A0',
         title: {
             style: {
-                color: '#AAA',
-                font: 'bold 12px Lucida Grande, Lucida Sans Unicode,' +
-                    ' Verdana, Arial, Helvetica, sans-serif'
+                color: '#CCC',
+                fontWeight: 'bold',
+                fontSize: '12px',
+                fontFamily: 'Trebuchet MS, Verdana, sans-serif'
+
             }
         }
-    },
+    } as Highcharts.XAxisOptions,
     yAxis: {
-        alternateGridColor: null,
-        minorTickInterval: null,
-        gridLineColor: 'rgba(255, 255, 255, .1)',
-        minorGridLineColor: 'rgba(255,255,255,0.07)',
-        lineWidth: 0,
-        tickWidth: 0,
+        gridLineColor: '#333333',
         labels: {
             style: {
-                color: '#999',
-                fontWeight: 'bold'
+                color: '#A0A0A0'
             }
         },
+        lineColor: '#A0A0A0',
+        minorTickInterval: null,
+        tickColor: '#A0A0A0',
+        tickWidth: 1,
         title: {
             style: {
-                color: '#AAA',
-                font: 'bold 12px Lucida Grande, Lucida Sans Unicode,' +
-                    ' Verdana, Arial, Helvetica, sans-serif'
+                color: '#CCC',
+                fontWeight: 'bold',
+                fontSize: '12px',
+                fontFamily: 'Trebuchet MS, Verdana, sans-serif'
             }
         }
-    },
-    legend: {
-        backgroundColor: 'rgba(48, 48, 48, 0.8)',
-        itemStyle: {
-            color: '#CCC'
-        },
-        itemHoverStyle: {
-            color: '#FFF'
-        },
-        itemHiddenStyle: {
-            color: '#333'
-        },
-        title: {
-            style: {
-                color: '#E0E0E0'
-            }
-        }
-    },
-    labels: {
-        style: {
-            color: '#CCC'
-        }
-    },
+    } as Highcharts.YAxisOptions,
     tooltip: {
-        backgroundColor: {
-            linearGradient: { x1: 0, y1: 0, x2: 0, y2: 1 },
-            stops: [
-                [0, 'rgba(96, 96, 96, .8)'],
-                [1, 'rgba(16, 16, 16, .8)']
-            ]
-        },
-        borderWidth: 0,
+        backgroundColor: 'rgba(0, 0, 0, 0.75)',
         style: {
-            color: '#FFF'
+            color: '#F0F0F0'
+        }
+    },
+    toolbar: {
+        itemStyle: {
+            color: 'silver'
         }
     },
     plotOptions: {
-        series: {
-            dataLabels: {
-                color: '#444'
-            },
-            nullColor: '#444444'
-        },
         line: {
             dataLabels: {
                 color: '#CCC'
@@ -147,11 +119,35 @@ Highcharts.theme = {
             lineColor: 'white'
         }
     },
-    toolbar: {
+    legend: {
+        backgroundColor: 'rgba(0, 0, 0, 0.5)',
         itemStyle: {
+            font: '9pt Trebuchet MS, Verdana, sans-serif',
+            color: '#A0A0A0'
+        },
+        itemHoverStyle: {
+            color: '#FFF'
+        },
+        itemHiddenStyle: {
+            color: '#444'
+        },
+        title: {
+            style: {
+                color: '#C0C0C0'
+            }
+        }
+    },
+    credits: {
+        style: {
+            color: '#666'
+        }
+    },
+    labels: {
+        style: {
             color: '#CCC'
         }
     },
+
     navigation: {
         buttonOptions: {
             symbolStroke: '#DDDDDD',
@@ -167,6 +163,7 @@ Highcharts.theme = {
             }
         }
     },
+
     // scroll charts
     rangeSelector: {
         buttonTheme: {
@@ -219,6 +216,7 @@ Highcharts.theme = {
             color: 'silver'
         }
     },
+
     navigator: {
         handles: {
             backgroundColor: '#666',
@@ -231,6 +229,7 @@ Highcharts.theme = {
             lineColor: '#A6C7ED'
         }
     },
+
     scrollbar: {
         barBackgroundColor: {
             linearGradient: { x1: 0, y1: 0, x2: 0, y2: 1 },
@@ -260,5 +259,6 @@ Highcharts.theme = {
         trackBorderColor: '#666'
     }
 };
+
 // Apply the theme
 setOptions(Highcharts.theme);
