+++ ts/Extensions/Themes/Grid.ts
@@ -9,10 +9,13 @@
  *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
  *
  * */
+
 'use strict';
-import Highcharts from '../Core/Globals.js';
-import U from '../Core/Utilities.js';
-var setOptions = U.setOptions;
+
+import Highcharts from '../../Core/Globals.js';
+import U from '../../Core/Utilities.js';
+const { setOptions } = U;
+
 Highcharts.theme = {
     colors: ['#058DC7', '#50B432', '#ED561B', '#DDDF00', '#24CBE5', '#64E572',
         '#FF9655', '#FFF263', '#6AF9C4'],
@@ -57,6 +60,7 @@ Highcharts.theme = {
                 fontWeight: 'bold',
                 fontSize: '12px',
                 fontFamily: 'Trebuchet MS, Verdana, sans-serif'
+
             }
         }
     },
@@ -85,6 +89,7 @@ Highcharts.theme = {
         itemStyle: {
             font: '9pt Trebuchet MS, Verdana, sans-serif',
             color: 'black'
+
         },
         itemHoverStyle: {
             color: '#039'
@@ -98,6 +103,7 @@ Highcharts.theme = {
             color: '#99b'
         }
     },
+
     navigation: {
         buttonOptions: {
             theme: {
@@ -106,5 +112,6 @@ Highcharts.theme = {
         }
     }
 };
+
 // Apply the theme
 setOptions(Highcharts.theme);
