+++ ts/Extensions/Themes/HighContrastLight.ts
@@ -13,10 +13,13 @@
  *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
  *
  * */
+
 'use strict';
-import Highcharts from '../Core/Globals.js';
-import U from '../Core/Utilities.js';
-var setOptions = U.setOptions;
+
+import Highcharts from '../../Core/Globals.js';
+import U from '../../Core/Utilities.js';
+const { setOptions } = U;
+
 Highcharts.theme = {
     colors: [
         '#5f98cf',
@@ -30,6 +33,7 @@ Highcharts.theme = {
         '#4f4a7a',
         '#b381b3'
     ],
+
     navigator: {
         series: {
             color: '#5f98cf',
@@ -37,5 +41,6 @@ Highcharts.theme = {
         }
     }
 };
+
 // Apply the theme
 setOptions(Highcharts.theme);
