+++ ts/Extensions/Themes/Skies.ts
@@ -12,8 +12,8 @@
 
 'use strict';
 
-import Highcharts from '../Core/Globals.js';
-import U from '../Core/Utilities.js';
+import Highcharts from '../../Core/Globals.js';
+import U from '../../Core/Utilities.js';
 const { setOptions } = U;
 
 Highcharts.theme = {
